Allow user to generate whatever type of work they want
